// src/app/types/dictionary.ts
export enum Dictionary {
  GLUE_SY = ' sy ',
  GLUE_AMBY = ' amby ',

  CUSTOM_ONE = 'iraika',

  ZERO = 'aotra',
  ONE = 'iray',
  TWO = 'roa',
  THREE = 'telo',
  FOUR = 'efatra',
  FIVE = 'dimy',
  SIX = 'enina',
  SEVEN = 'fito',
  EIGHT = 'valo',
  NINE = 'sivy',

  TEN = 'folo',
  TWENTY = 'roapolo',
  THIRTY = 'telopolo',
  FORTY = 'efapolo',
  FIFTY = 'dimampolo',
  SIXTY = 'enimpolo',
  SEVENTY = 'fitopolo',
  EIGHTY = 'valopolo',
  NINETY = 'sivifolo',

  HUNDRED = 'zato',
  TWO_HUNDRED = 'roanjato',
  THREE_HUNDRED = 'telonjato',
  FOUR_HUNDRED = 'efajato',
  FIVE_HUNDRED = 'dimanjato',
  SIX_HUNDRED = 'enin-jato',
  SEVEN_HUNDRED = 'fitonjato',
  EIGHT_HUNDRED = 'valonjato',
  NINE_HUNDRED = 'sivinjato',

  THOUSAND = 'arivo',
  TWO_THOUSAND = 'roa arivo',
  THREE_THOUSAND = 'telo arivo',
  FOUR_THOUSAND = 'efatra arivo',
  FIVE_THOUSAND = 'dimy arivo',
  SIX_THOUSAND = 'enina arivo',
  SEVEN_THOUSAND = 'fito arivo',
  EIGHT_THOUSAND = 'valo arivo',
  NINE_THOUSAND = 'sivy arivo',

  TEN_THOUSAND = 'iray alina',
  TWENTY_THOUSAND = 'roa alina',
  THIRTY_THOUSAND = 'telo alina',
  FORTY_THOUSAND = 'efatra alina',
  FIFTY_THOUSAND = 'dimy alina',
  SIXTY_THOUSAND = 'enina alina',
  SEVENTY_THOUSAND = 'fito alina',
  EIGHTY_THOUSAND = 'valo alina',
  NINETY_THOUSAND = 'sivy alina',

  ONE_HUNDRED_THOUSAND = 'iray hetsy',
  TWO_HUNDRED_THOUSAND = 'roa hetsy',
  THREE_HUNDRED_THOUSAND = 'telo hetsy',
  FOUR_HUNDRED_THOUSAND = 'efatra hetsy',
  FIVE_HUNDRED_THOUSAND = 'dimy hetsy',
  SIX_HUNDRED_THOUSAND = 'enina hetsy',
  SEVEN_HUNDRED_THOUSAND = 'fito hetsy',
  EIGHT_HUNDRED_THOUSAND = 'valo hetsy',
  NINE_HUNDRED_THOUSAND = 'sivy hetsy',

  MILLION = 'iray tapitrisa',
  TWO_MILLION = 'roa tapitrisa',
  THREE_MILLION = 'telo tapitrisa',
  FOUR_MILLION = 'efatra tapitrisa',
  FIVE_MILLION = 'dimy tapitrisa',
  SIX_MILLION = 'enina tapitrisa',
  SEVEN_MILLION = 'fito tapitrisa',
  EIGHT_MILLION = 'valo tapitrisa',
  NINE_MILLION = 'sivy tapitrisa',

  BILLION = 'iray lavitrisa',
  TWO_BILLION = 'roa lavitrisa',
  THREE_BILLION = 'telo lavitrisa',
  FOUR_BILLION = 'efatra lavitrisa',
  FIVE_BILLION = 'dimy lavitrisa',
  SIX_BILLION = 'enina lavitrisa',
  SEVEN_BILLION = 'fito lavitrisa',
  EIGHT_BILLION = 'valo lavitrisa',
  NINE_BILLION = 'sivy lavitrisa',
}